requirements.txt
docopt==0.6.2
requests==2.7.0
fluent-logger
